% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listeners.R
\name{hot_add_listener}
\alias{hot_add_listener}
\title{Set event listeners on the hot object}
\usage{
hot_add_listener(hot, hook, jsFunction, local_or_global = "local")
}
\arguments{
\item{hook}{The name of the handsontable hook you're attaching the listener to.}

\item{jsFunction}{Your javascript function passed as a string.}
}
\description{
Set event listeners on the hot object
}
