% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhot.R
\name{hot_cell}
\alias{hot_cell}
\title{Set cell-level settings.}
\usage{
hot_cell(hot, rowId = NULL, columnId = NULL, readOnly = NULL, comment = NULL)
}
\arguments{
\item{comment}{Either a character matrix or a character vector. If character matrix then rowId and columnId are disregarded. Comments are passed at the cell-level to the cell object.}
}
\description{
Set cell-level settings.
}
