% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhot.R
\name{hot_column}
\alias{hot_column}
\title{Set column-level settings.}
\usage{
hot_column(hot, columnId, returnType = NULL, allowInvalid = NULL,
  copyable = NULL, copyPaste = NULL, correctFormat = NULL,
  dateFormat = NULL, defaultDate = NULL, editor = NULL,
  filter = NULL, filteringCaseSensitive = NULL, label = NULL,
  numericFormat = NULL, placeholder = NULL, readOnly = NULL,
  renderer = NULL, selectOptions = NULL, skipColumnOnPaste = NULL,
  source = NULL, strict = NULL, title = NULL, trimDropdown = NULL,
  trimWhitespace = NULL, type = NULL, validator = NULL,
  visibleRows = NULL, wordWrap = NULL, halign = NULL,
  valign = NULL, ...)
}
\arguments{
\item{hot}{The Handsontable object created by rhot().}

\item{columnId}{Vector to specify which columns for the provided settings. Can be numeric vector of column indexes or character vector of column names.}

\item{returnType}{Sets the R-class that will be used to deserialize data back into R. By default these are set by the class of the columns in the data.frame provided to rhot().}

\item{renderer}{default options: 'autocomplete', 'checkbox', 'html', 'numeric', 'password', 'text' or custRenderer}

\item{type}{Character vector defining the type for each column. This is the preferred way of setting the editor, renderer, and validator (ERV) for a column when using predefined column types. To further customize your table, supply the type then override any or all of the three ERV settings.}

\item{validator}{Character vector with names of validator to be used for the column. Setting validator explicitly will override the validator that is implied by setting the column type. Handsontable provides 'autocomplete','date','dropdown', 'numeric', 'time' as prebuilt validators. See details for explanation of each. For further customization you can register your own custom validator by using \code{\link{rhot::validator_register_custom}} and then passing that validator's name to this parameter.}

\item{...}{Additional settings passed to the column object.}
}
\description{
Pass settings to the handsontable object that are column specific. The function is vectorized over columns and the settings. This means that you can set the same settings for different columns and different settings for different columns in the same call.
}
\details{
## Type, Editor, Renderer, Validator

## Validator
### Standard options
These are the options provided by Handsontable.
autocomplete
date
dropdown
numeric
time
custom validator: see rhot::validator_register_custom()
### Custom Validator
}
