% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhot.R
\name{hot_listener}
\alias{hot_listener}
\title{Set event listeners on the hot object}
\usage{
hot_listener(hot, ...)
}
\arguments{
\item{...}{The event listeners.

Pass event listeners to the hot library. Name the parameter as the listener you are setting.
Each parameter's content is the event listener function as a string.

There are pre-set event handlers. This function is intended to allow the user to add more event
listeners but if you really want to turn off a pre-set handler then pass that listener as
an empty string and it'll be overwritten.
Default handlers: afterChange}
}
\description{
Set event listeners on the hot object
}
