% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhot.R
\name{hot_menu}
\alias{hot_menu}
\title{Configure table's context menu.}
\usage{
hot_menu(hot, menu = TRUE, allowInsertColumn = NULL)
}
\arguments{
\item{menu}{TRUE or FALSE to add a menu to the table. Defaults is TRUE in this function but note that the table itself defaults the context menu to FALSE.}

\item{moreparams}{}
}
\description{
A context menu can be configured and added to the table through this function.
}
