% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhot.R
\name{hot_register_editor}
\alias{hot_register_editor}
\title{Register a custom editor with Handsontable}
\usage{
hot_register_editor(hot, name, jsFunction)
}
\arguments{
\item{name}{Name your editor}

\item{jsFunction}{Your javascript editor function passed as a string.

TODO Add details for what is expected of a validator function.}
}
\description{
Register a custom editor with Handsontable
}
