% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhot.R
\name{hot_register_renderer}
\alias{hot_register_renderer}
\title{Register a custom cell renderer}
\usage{
hot_register_renderer(hot, name, jsFunction)
}
\arguments{
\item{name}{Name your renderer.}

\item{jsFunction}{Your javascript renderer function passed as a string.

TODO Add details for what is expected of a renderer function.}
}
\description{
Register a custom cell renderer
}
