% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhot.R
\name{hot_register_validator}
\alias{hot_register_validator}
\title{Register a custom cell validator}
\usage{
hot_register_validator(hot, name, jsFunction)
}
\arguments{
\item{name}{Name your validator.}

\item{jsFunction}{Your javascript validator function passed as a string.

TODO Add details for what is expected of a validator function.}
}
\description{
Register a custom cell validator
}
