% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhot.R
\name{hot_style}
\alias{hot_style}
\title{Set styles for the table.}
\usage{
hot_style(hot, someClassSettings = NULL)
}
\arguments{
\item{hot}{}
}
\description{
Change the defualt classes that are used for styling the table. Also, add custom borders to the table.
}
