% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhot.R
\name{hot_table}
\alias{hot_table}
\title{Set table-level settings.}
\usage{
hot_table(
  hot,
  autoColumnSize = NULL,
  autoRowSize = NULL,
  autoWrapCol = NULL,
  autoWrapRow = NULL,
  colHeaders = TRUE,
  columnHeaderHeight = NULL,
  colWidths = NULL,
  dragToScroll = NULL,
  enterMoves = NULL,
  fixedColumnsLeft = NULL,
  fixedRowsBottom = NULL,
  fixedRowsTop = NULL,
  fragmentSelection = NULL,
  language = NULL,
  manualColumnFreeze = NULL,
  manualColumnMove = NULL,
  manualColumnResize = NULL,
  manualRowMove = NULL,
  manualRowResize = NULL,
  maxCols = NULL,
  maxRows = NULL,
  mergeCells = NULL,
  minCols = NULL,
  minRows = NULL,
  observeDOMVisibility = NULL,
  outsideClickDeselects = NULL,
  rowHeaders = NULL,
  rowHeaderWidth = NULL,
  rowHeights = NULL,
  search = NULL,
  selectionMode = NULL,
  stretchH = "none",
  tabMoves = NULL,
  titles = NULL,
  undo = NULL,
  validate_upon_creation = NULL,
  ...
)
}
\arguments{
\item{colHeaders}{Logical, by default colHeaders is set to TRUE and
rhot sets the column headers using the colnames of the provided dataframe.
To set the column headers you should use the title parameter.
To set a single column's header you should use the title parameter of hot_column().}

\item{titles}{Set all column titles at once. Must be character vector of length one or ncol(data).}

\item{validate_upon_creation}{something to validate the initial data}

\item{...}{passed onto handsontable as table-level settings.}
}
\description{
Set table-level settings.
}
