% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators_and_renderers.R
\name{register_renderer_custom}
\alias{register_renderer_custom}
\title{Register a custom validator with Handsontable}
\usage{
register_renderer_custom(hot, name, jsFunction)
}
\description{
Register a custom validator with Handsontable
}
