% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators_and_renderers.R
\name{register_renderer_heatmap}
\alias{register_renderer_heatmap}
\title{Register a custom validator with Handsontable}
\usage{
register_renderer_heatmap(
  hot,
  min_value = 0,
  max_value = 1,
  min_color = "#000000",
  max_color = "#ff0000"
)
}
\description{
Register a custom validator with Handsontable
}
