% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators_and_renderers.R
\name{register_validator_character}
\alias{register_validator_character}
\title{Register a custom validator with Handsontable}
\usage{
register_validator_character(hot, name)
}
\description{
Register a custom validator with Handsontable
}
