% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators_and_renderers.R
\name{register_validator_numeric}
\alias{register_validator_numeric}
\title{Register a custom validator with Handsontable}
\usage{
register_validator_numeric(
  hot,
  name,
  max_value = NULL,
  min_value = NULL,
  inclusive_above = TRUE,
  inclusive_below = TRUE,
  integers_only = FALSE
)
}
\description{
Register a custom validator with Handsontable
}
