% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhot.R
\name{rhot}
\alias{rhot}
\title{<Add Title>}
\usage{
rhot(elementId, data, width = "auto", height = "auto",
  overflow = "hidden")
}
\arguments{
\item{elementId}{The id given to the widget.}

\item{data}{The data.frame to fill the table.}

\item{width, height}{Dimensions for the widget's container.}

\item{overflow}{This css value is passed on to the widget's container. Its not recommended to change this.}
}
\description{
<Add Description>
}
